Rename @monorepolint/cli to monorepolint.
This makes it so that `packages/all` is purely a local build sink.
Instead, `packages/cli` is now used as the public facing package. The
reason being this allows for the actual binary for the cli to be in a
predictable location regardless of the npm client being used.

Previously, there was a difference of location when using yarn vs yarn
workspaces. The latter would always hoist the binary while the former
did not.