Ensure Context.addError calls have non-null file arguments
The `Context` interface defines an `addError` method accepting a
nullable `file` argument. The `PackageContext` class implements this
with a method requiring the `file` argument. This is a type mismatch.

A new `AddErrorOptions` interface is added that extends `Failure`, so
callers know that they're required to pass in a defined `file` to
`PackageContext.addError`.